Prepare for router
