+++ src/apps/settings/SettingsModal.tsx
@@ -7,6 +7,10 @@ import { OpenAIAdvancedSettings } from '@/modules/openai/OpenAIAdvancedSettings'
 import { OpenAISettings } from '@/modules/openai/OpenAISettings';
 import { ProdiaSettings } from '@/modules/prodia/ProdiaSettings';
 import { SearchSettings } from '@/modules/search/SearchSettings';
+import { isValidOpenAIApiKey, requireUserKeyOpenAI } from '@/modules/openai/openai.client';
+
+import { useSettingsStore } from '@/common/state/store-settings';
+import { useUIStore } from '@/common/state/store-ui';
 
 import { UISettings } from './UISettings';
 
@@ -14,13 +18,20 @@ import { UISettings } from './UISettings';
 /**
  * Component that allows the User to modify the application settings,
  * persisted on the client via localStorage.
- *
- * @param {boolean} open Whether the Settings modal is open
- * @param {() => void} onClose Call this to close the dialog from outside
  */
-export function SettingsModal({ open, onClose }: { open: boolean, onClose: () => void; }) {
+export function SettingsModal() {
+  // external state
+  const { settingsOpen, closeSettings } = useUIStore();
+  const apiKey = useSettingsStore(state => state.apiKey);
+
+  // show the Settings Dialog at startup if the API key is required but not set
+  React.useEffect(() => {
+    if (requireUserKeyOpenAI && !isValidOpenAIApiKey(apiKey))
+      useUIStore.getState().openSettings();
+  }, [apiKey]);
+
   return (
-    <Modal open={open} onClose={onClose}>
+    <Modal open={settingsOpen} onClose={closeSettings}>
       <ModalOverflow>
         <ModalDialog sx={{ maxWidth: 500, display: 'flex', p: { xs: 1, sm: 2, lg: '20px' } }}>
 
@@ -40,7 +51,7 @@ export function SettingsModal({ open, onClose }: { open: boolean, onClose: () =>
           <OpenAIAdvancedSettings />
 
           <Box sx={{ mt: 4, display: 'flex', justifyContent: 'flex-end' }}>
-            <Button variant='solid' onClick={onClose}>
+            <Button variant='solid' onClick={closeSettings}>
               Close
             </Button>
           </Box>
